Refactor Multivariate modules
