+++ src/ZkFold/Base/Algebra/Polynomials/Multivariate/Monomial.hs
@@ -5,20 +5,32 @@
 module ZkFold.Base.Algebra.Polynomials.Multivariate.Monomial
     ( M(..)
     , Monomial
+    , Monomial'
+    , monomial
     , Variable
+    , MonomialAny
+    , MonomialRepAny
+    , MonomialRepBoundedDegree
+    , MonomialBoundedDegree
+    , evalMapM
+    , evalVectorM
+    , mapVar
+    , mapVarMonomial
     ) where
 
 import           Control.DeepSeq                 (NFData)
 import           Data.Aeson                      (FromJSON, ToJSON)
 import           Data.List                       (intercalate)
-import           Data.Map.Strict                 (Map, differenceWith, empty, unionWith)
+import           Data.Map.Strict                 (Map, differenceWith, filter, foldrWithKey, fromListWith, mapKeys, empty, unionWith)
 import qualified Data.Map.Strict                 as Map
 import           GHC.Generics                    (Generic)
 import           GHC.IsList                      (IsList (..))
 import           Numeric.Natural                 (Natural)
-import           Prelude                         hiding (Num (..), drop, lcm, length, sum, take, (!!), (/))
+import           Prelude                         hiding (Num (..), drop, filter, lcm, length, sum, take, (^), (!!), (/))
 import           Test.QuickCheck                 (Arbitrary (..))
 
+import           ZkFold.Prelude                  (elemIndex)
+import           ZkFold.Base.Data.Vector         (Vector (..))
 import           ZkFold.Base.Algebra.Basic.Class
 
 type Variable i = Ord i
@@ -28,6 +40,26 @@ type Monomial i j = (Variable i, Ord j, Semiring j)
 -- | Monomial type
 newtype M i j m = M m
     deriving (Generic, NFData, FromJSON, ToJSON)
+
+-- | Most general type for a multivariate monomial
+type Monomial' = M Natural Natural (Map Natural Natural)
+
+type MonomialRepAny = Map Integer Integer
+
+type MonomialRepBoundedDegree i d = Vector d (i, Bool)
+
+type MonomialAny = M Integer Integer MonomialRepAny
+
+type MonomialBoundedDegree i d = M i Bool (MonomialRepBoundedDegree i d)
+
+instance Ord i => IsList (M i j (Map i j)) where
+    type Item (M i j (Map i j)) = (i, j)
+    toList (M m) = toList m
+    fromList m = M $ fromList m
+
+-- | Monomial constructor
+monomial :: Monomial i j => Map i j -> M i j (Map i j)
+monomial = M . filter (/= zero)
 
 instance (Show i, Show j, Monomial i j) => Show (M i j (Map i j)) where
     show (M m) = intercalate "∙" . map showVar $ toList m
@@ -68,3 +100,27 @@ instance (Monomial i j, Ring j) => MultiplicativeGroup (M i j (Map i j)) where
 
     M l / M r = M $ differenceWith f l r
         where f a b = if a == b then Nothing else Just (a - b)
+
+evalMapM :: forall i j b .
+    MultiplicativeMonoid b =>
+    Exponent b j =>
+    (i -> b) -> M i j (Map i j) -> b
+evalMapM f (M m) =
+    foldrWithKey (\i j x -> (f i ^ j) * x) (one @b) m
+
+evalVectorM :: forall i j b d .
+    Monomial i j =>
+    MultiplicativeMonoid b =>
+    Exponent b j =>
+    (i -> b) -> M i j (Vector d (i, Bool)) -> b
+evalVectorM f (M (Vector v)) =
+    evalMapM f . M . fromListWith (+)
+        $ foldr (\(i, x) xs -> if x then (i, one @j) : xs else xs) [] v
+
+mapVar :: [Natural] -> Natural -> Natural
+mapVar vars x = case x `elemIndex` vars of
+    Just i  -> i
+    Nothing -> error "mapVar: something went wrong"
+
+mapVarMonomial :: [Natural] -> Monomial' -> Monomial'
+mapVarMonomial vars (M as) = M $ mapKeys (mapVar vars) as
