+++ zkfold-base.cabal
@@ -93,7 +93,6 @@ library
       ZkFold.Base.Algebra.Polynomials.Multivariate
       ZkFold.Base.Algebra.Polynomials.Multivariate.Monomial
       ZkFold.Base.Algebra.Polynomials.Multivariate.Polynomial
-      ZkFold.Base.Algebra.Polynomials.Multivariate.Set
       ZkFold.Base.Algebra.Polynomials.Multivariate.Substitution
       ZkFold.Base.Algebra.Polynomials.Univariate
       ZkFold.Base.Data.ByteString
